Separating eth_filter
