+++ rpc/src/v1/impls/mod.rs
@@ -27,6 +27,7 @@ macro_rules! take_weak {
 
 mod web3;
 mod eth;
+mod eth_filter;
 mod net;
 mod personal;
 mod ethcore;
@@ -34,7 +35,8 @@ mod traces;
 mod rpc;
 
 pub use self::web3::Web3Client;
-pub use self::eth::{EthClient, EthFilterClient, EthSigningUnsafeClient};
+pub use self::eth::{EthClient, EthSigningUnsafeClient};
+pub use self::eth_filter::EthFilterClient;
 pub use self::net::NetClient;
 pub use self::personal::PersonalClient;
 pub use self::ethcore::EthcoreClient;
